[Nilima]AAded Data

+++ Address_book-Ao/Address_book-Ao/AddressBookModel.cs
@@ -16,7 +16,7 @@ namespace Address_book_Ao
         public string Address { get; set; }
         public string City { get; set; }
         public string State { get; set; }
-        public long Zip { get; set; }
+        public int  Zip { get; set; }
         public long PhoneNumber { get; set; }
         public string EmailId { get; set; }
         public string AddressBookType { get; set; }
